web-forms (Vue UI): break out upload image preview component
… leaving more of the base component suitable for generic upload concerns
